Use JS for config files
