+++ web/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../tsconfig.json",
   "references": [{ "path": "../shared" }, { "path": "../schema" }],
   "compilerOptions": {
-    "module": "esnext",
+    "module": "commonjs",
     "typeRoots": ["../shared/src/types", "../node_modules/@types", "src/types"],
     "jsx": "react",
     "rootDir": ".",
